etl_waterlevel/requirements.txt
pandas==2.1.3
requests==2.31.0
urllib3==2.0.7
psycopg2-binary==2.9.9